Moved some connection management related methods to ArJdbc::Abstract::ConnectionManagement
