+++ lib/arjdbc/abstract/connection_management.rb
@@ -1,0 +1,22 @@
+module ArJdbc
+  module Abstract
+    module ConnectionManagement
+
+      def active?
+        @connection.active?
+      end
+
+      def disconnect!
+        super
+        @connection.disconnect!
+      end
+
+      def reconnect!
+        super
+        @connection.reconnect! # handles adapter.configure_connection
+        @connection
+      end
+
+    end
+  end
+end
